split benchmarks + surrounding utils into two files
